HomogeneousTransformations.jl

HomogeneousTransformations.jl provides basic functions for creating and using homogeneous transformation matrices to perform 3D translations and rotations.

References:

2006SpongRobotModelingAndControl

GitHub source

HomogeneousTransformations.HMatrix — Type

HMatrix wraps a SMatrix{4,4}, enforcing

defined size of 4x4

unit columns in rotation matrix [1:3,1:3]

0001 last row

data::StaticArraysCore.MMatrix

HomogeneousTransformations.Eye — Method

Eye() -> HMatrix

Returns an Identity matrix.

HomogeneousTransformations.Rtk — Method

Rtk(th, k) -> HMatrix

Create a rotation of angle th, in radians, about and aribrary axis k.

HomogeneousTransformations.Rx — Method

Rx(a) -> HMatrix

Create a rotation matrix of angle a about the X axis, in radians.

HomogeneousTransformations.Ry — Method

Ry(b) -> HMatrix

Create a rotation matrix of angle b about the Y axis, in radians.

HomogeneousTransformations.Rz — Method

Rz(c) -> HMatrix

Create a rotation matrix of angle c about the Z axis, in radians.

HomogeneousTransformations.Tx — Method

Tx(a) -> HMatrix

Create a transformation matrix of distance a along the X axis.

HomogeneousTransformations.Ty — Method

Ty(b) -> HMatrix

Create a transformation matrix of distance b along the Y axis.

HomogeneousTransformations.Tz — Method

Tz(c) -> HMatrix

Create a transformation matrix of distance c along the Z axis.